requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.55.0
requests==2.32.3
streamlit
pandas
plotly